Split implementation out into files
