+++ src/test/index.ts
@@ -1,3 +1,17 @@
+/**
+ * @license
+ * Copyright (c) 2016 The Polymer Project Authors. All rights reserved.
+ * This code may only be used under the BSD style license found at
+ * http://polymer.github.io/LICENSE.txt
+ * The complete set of authors may be found at
+ * http://polymer.github.io/AUTHORS.txt
+ * The complete set of contributors may be found at
+ * http://polymer.github.io/CONTRIBUTORS.txt
+ * Code distributed by Google as part of the polymer project is also
+ * subject to an additional IP rights grant found at
+ * http://polymer.github.io/PATENTS.txt
+ */
+
 /// <reference path="../../node_modules/@types/mocha/index.d.ts" />
 
 import {assert} from 'chai';
@@ -6,14 +20,13 @@ import * as path from 'path';
 import {Analyzer} from 'polymer-analyzer';
 import {FSUrlLoader} from 'polymer-analyzer/lib/url-loader/fs-url-loader';
 
-import {UpgradePassRegistry, Upgrader} from '../index';
+import {Upgrader} from '../upgrader';
 
 suite('upgrade()', function() {
   const fixturesDir = path.join(__dirname, 'fixtures');
   const dirnames = fs.readdirSync(fixturesDir);
   for (const dirname of dirnames) {
-    const upgrader =
-        new Upgrader({upgradePasses: UpgradePassRegistry.allPassCodes});
+    const upgrader = new Upgrader({upgradePasses: ['all']});
     test(`upgrades ${dirname}`, async() => {
       const fullDir = path.join(fixturesDir, dirname);
       const beforeDir = path.join(fullDir, 'before');
