+++ src/upgrade-pass.ts
@@ -1,0 +1,29 @@
+/**
+ * @license
+ * Copyright (c) 2016 The Polymer Project Authors. All rights reserved.
+ * This code may only be used under the BSD style license found at
+ * http://polymer.github.io/LICENSE.txt
+ * The complete set of authors may be found at
+ * http://polymer.github.io/AUTHORS.txt
+ * The complete set of contributors may be found at
+ * http://polymer.github.io/CONTRIBUTORS.txt
+ * Code distributed by Google as part of the polymer project is also
+ * subject to an additional IP rights grant found at
+ * http://polymer.github.io/PATENTS.txt
+ */
+import {ParsedDocument} from 'polymer-analyzer/lib/parser/document';
+
+export abstract class UpgradePass<T extends ParsedDocument<any, any>> {
+  code: string;
+  documentClazz: any;
+
+  constructor() {}
+
+  canUpgrade(document: T) { return document instanceof this.documentClazz; }
+
+  abstract upgrade(document: T): void;
+}
+
+export class UpgradePassCollection {
+  constructor(public code: string, public passes: string[]) {}
+}
